Got Character Moving!
I only had to delete all of the old broken blueprints and manually rebuild the logic! This is a win tho since I won't have to rewrite all of the Boilerplate GAS setup code and work directly off of Tranek's documentation project. This will allow me to get directly into content creation and tweaking the c++ as needed rather than writing a ton of it upfront.