Debug for Ability Slot Assigning in Character
Created several keyboard events that simulate assigning abilities to ability slots.
Pressing Z, X, or C key will initiate assigning Test Ability Alpha, Beta, and Charlie respectively.
Pressing 1, 2, or 3 key will activate the ability in the respective slot unless ability assigning has been initiated. In that case, it would assign the ability to that slot instead of activating it.
Right now all these abilities do is print a debug log, but in the future this will be pivotal to character customization and gameplay.