Worked on Ammo
Added Ammo UI
Moved HUD UI to PlayerController
Reorganized some files
Added CapacityReserve and MaxCapacityReserve attributes to a bunch of c++ files
The weapon actor will handle the current ammo that each gun is using while an attribute is made to track the reserve ammo that the player has. This may be split up in the future to multiple ammo types.